Update cache.py
Related to issue https://github.com/edoburu/django-parler/issues/49

Removed the "long" from the format of master_id, so now can support not integer values for primary_keys.

+++ parler/cache.py
@@ -57,7 +57,7 @@ def get_translation_cache_key(translated_model, master_id, language_code):
     """
     # Always cache the entire object, as this already produces
     # a lot of queries. Don't go for caching individual fields.
-    return 'parler.{0}.{1}.{2}.{3}'.format(translated_model._meta.app_label, translated_model.__name__, long(master_id), language_code)
+    return 'parler.{0}.{1}.{2}.{3}'.format(translated_model._meta.app_label, translated_model.__name__, master_id, language_code)
 
 
 def get_cached_translation(instance, language_code=None, related_name=None, use_fallback=False):
